+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.2.0
+
+- Added domain allowlist/blocklist feature to control where the extension operates (thanks [@Aptcoder](https://github.com/Aptcoder)!)
+  - Users can specify domains where the extension should or should not run based on allow or block mode (default is allow)
+  - Supports wildcard domains (e.g. *.example.com)
+  - Supports specific overrides (e.g. if google.com is blocked, google.com/maps can be allowed)
+  - Adds color-coded status indicator showing whether name replacement is enabled or disabled on the current site
+  - Updates extension icon to reflect the current site's allowlist/blocklist status
+
 ## v2.1.0
 
 - Added keyboard shortcut functionality for quickly enabling/disabling the extension